FrameBuffer: Cast size x,y to size_t to avoid conversion warning

+++ src/FrameBuffers.hpp
@@ -30,7 +30,7 @@ public:
         pixels.resize(height, std::vector<Color>(width, defaultColor));
     }
     FrameBuffer(const Vec2& size, const Color& defaultColor) :
-        FrameBuffer(size.x, size.y, defaultColor)
+        FrameBuffer(static_cast<size_t>(size.x), static_cast<size_t>(size.y), defaultColor)
     {}
 
     constexpr size_t getWidth()                   const noexcept { return width;            }
